ZNA-6197 - [IMD] - Landscape Player - Login screen appears for few seconds and user navigates to Portrait player on tapping Login button in Landscape player
Issue is also observed when user taps on Add To Watchlist in landscape mode.
When changing orientation mode then a lot of operations on currently shown viewcontrollers are performed(dismiss, remove from super, add as child a.t.c)
One of the options to fix it is just to start showing login screen after orientation changes has been performed.